backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
openai
passlib[bcrypt]
httpx
pytest
python-dotenv
scikit-learn
pandas
numpy
google-cloud-vision
pytesseract
pillow
apscheduler
google-generativeai
